backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Neo4j
neo4j==5.14.1

# ChromaDB and Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# LangChain and AI
langchain==0.0.335
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.19.4

# HTTP Client
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Security Scanning Tools (Python bindings)
python-nmap==0.7.1

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# WebSockets
websockets==12.0

# Task Scheduling
apscheduler==3.10.4
celery==5.3.4

# Email
python-mail==0.1.0
aiosmtplib==3.0.1
jinja2==3.1.2

# OpenVAS Integration
python-gvm==23.11.0
gvm-tools==23.11.0

# Nessus Integration
tenable-io==1.4.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Monitoring and Logging
python-json-logger==2.0.7
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pylint==3.0.3

# Development
ipython==8.18.1
rich==13.7.0
